livekit_agent/requirements.txt
livekit-agents[turn-detector,silero,codecs]~=1.3.10
livekit-plugins-noise-cancellation~=0.2
python-dotenv~=1.0.1
aiohttp~=3.10.11
